float

+++ youtube-dl.py
@@ -69,13 +69,13 @@ def info(id):
 def frame(id):
     # params = list(request.args.items())
     info = get_info(id)
-    duration = int(info["duration"])
+    duration = float(info["duration"])
     videoUrl = get_url(info)["video"]
 
     if not videoUrl:
         return create_response("no video")
 
-    ss = min(int(request.args["ss"]), duration) if "ss" in request.args else 0
+    ss = min(float(request.args["ss"]), duration) if "ss" in request.args else 0
     p = subprocess.run([
         ffmpeg,
         "-ss", str(ss),
